Implement core scraping functionality
This thing scrapes now! Still a long road to go in terms of ensuring it
writes to a DB and the filters are properly configurable, but this is a
huge leap forward. A number of functions were swapped from sync to async
to facilitate scraping without blocking essential bot functions, so
that's why a fair number of files needed edits here.