Restructure config objects, continue modal implementation
The reordering of the config objects is basically a refactor, though it
was done to facilitate efficiency in saving configs locally via the
configure command. The plan is to pop up a button for the user to apply
their changes, wherein a press will save the new configuration locally
to persist between sessions.

I was inclined to think of the button as a step back -- why beg for an
extra click? After some thought, I feel it's necessary to reduce the
frequency of file writes. It's unlikely the user would want to keep the
menu up after making their changes, anyway!